+++ packages/sit-onyx/src/components/OnyxCheckbox/OnyxCheckbox.vue
@@ -46,6 +46,7 @@ const { densityClass } = useDensity(props);
         v-else
         v-model="isChecked"
         :aria-label="props.hideLabel ? props.label : undefined"
+        :title="props.hideLabel ? props.label : undefined"
         class="onyx-checkbox__input"
         :class="{
           'onyx-checkbox__input--touched': isTouched,
@@ -58,13 +59,24 @@ const { densityClass } = useDensity(props);
       />
     </div>
 
-    <p
-      v-if="!props.hideLabel"
-      class="onyx-checkbox__label"
-      :class="[`onyx-truncation-${props.truncation}`, requiredMarkerClass]"
-    >
-      {{ props.label }}
-    </p>
+    <template v-if="!props.hideLabel">
+      <p
+        class="onyx-checkbox__label"
+        :class="[
+          `onyx-truncation-${props.truncation}`,
+          // shows the required marker inline for multiline labels
+          props.truncation === 'multiline' ? requiredMarkerClass : undefined,
+        ]"
+      >
+        {{ props.label }}
+      </p>
+      <!-- shows the required marker fixed on the right for truncated labels -->
+      <div
+        v-if="props.truncation === 'ellipsis'"
+        class="onyx-checkbox__marker"
+        :class="[requiredMarkerClass]"
+      ></div>
+    </template>
   </label>
 </template>
 
@@ -216,6 +228,10 @@ const { densityClass } = useDensity(props);
   &__label {
     display: inline-block;
     margin: 0;
+  }
+
+  &__label,
+  &__marker {
     padding: $label-padding 0;
     font-size: 1rem;
     line-height: 1.5rem;