+++ packages/sit-onyx/src/components/OnyxCheckbox/types.ts
@@ -26,7 +26,7 @@ export type OnyxCheckboxProps = DensityProp &
      */
     loading?: boolean;
     /**
-     * If `true`, the label will be visually hidden.
+     * If `true`, the label will be visually hidden and the `title` attribute will be set.
      * For accessibility / screen readers, the aria-label will still be set.
      */
     hideLabel?: boolean;