+++ packages/sit-onyx/src/components/OnyxIconButton/OnyxIconButton.vue
@@ -27,6 +27,7 @@ const emit = defineEmits<{
   <button
     class="onyx-icon-button"
     :aria-label="props.label"
+    :title="props.label"
     :class="[
       `onyx-icon-button--${props.variation}`,
       { 'onyx-icon-button--loading': props.loading },