+++ packages/sit-onyx/src/components/OnyxSwitch/OnyxSwitch.vue
@@ -69,12 +69,21 @@ watch(
     <OnyxSkeleton v-if="!props.hideLabel" class="onyx-switch-skeleton__label" />
   </div>
 
-  <label v-else class="onyx-switch" :class="[requiredTypeClass, densityClass]">
+  <label
+    v-else
+    class="onyx-switch"
+    :class="[requiredTypeClass, densityClass]"
+    :title="props.hideLabel ? props.label : undefined"
+  >
+    <!-- Linter incorrectly finds an error. For a native `input` the `aria-checked` is not necessary. There is an open issue about it: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/issues/932  -->
+    <!-- eslint-disable vuejs-accessibility/role-has-required-aria-props -->
+    <!-- TODO: disable can be removed when https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/pull/1071 was released -->
     <input
       ref="inputElement"
       v-model="isChecked"
-      :class="{ 'onyx-switch__input': true, 'onyx-switch__loading': props.loading }"
       type="checkbox"
+      role="switch"
+      :class="{ 'onyx-switch__input': true, 'onyx-switch__loading': props.loading }"
       :aria-label="props.hideLabel ? props.label : undefined"
       :disabled="props.disabled || props.loading"
       :required="props.required"