+++ packages/sit-onyx/src/components/OnyxSwitch/types.ts
@@ -25,7 +25,7 @@ export type OnyxSwitchProps = DensityProp &
      */
     errorMessage?: string;
     /**
-     * If `true`, the label will be visually hidden.
+     * If `true`, the label will be visually hidden and the `title` attribute will be set.
      * For accessibility / screen readers, the aria-label will still be set.
      */
     hideLabel?: boolean;