+++ packages/sit-onyx/src/components/OnyxTooltip/OnyxTooltip.stories.ts
@@ -3,6 +3,8 @@ import { TOOLTIP_TRIGGERS } from "@sit-onyx/headless";
 import circleInformation from "@sit-onyx/icons/circle-information.svg?raw";
 import { defineStorybookActionsAndVModels } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
+import { h } from "vue";
+import OnyxButton from "../OnyxButton/OnyxButton.vue";
 import OnyxTooltip from "./OnyxTooltip.vue";
 
 /**
@@ -16,6 +18,9 @@ const meta: Meta<typeof OnyxTooltip> = {
     argTypes: {
       icon: defineIconSelectArgType(),
       default: {
+        control: { disabled: true },
+      },
+      tooltip: {
         control: { disabled: true },
       },
       open: {
@@ -43,7 +48,7 @@ type Story = StoryObj<typeof OnyxTooltip>;
 export const Default = {
   args: {
     text: "Tooltip text",
-    default: () => "Slot content goes here",
+    default: () => h(OnyxButton, { label: "Slot content goes here" }),
     icon: circleInformation,
     open: true,
   },
@@ -56,7 +61,7 @@ export const Default = {
 export const Hover = {
   args: {
     ...Default.args,
-    default: () => "Hover me to show the tooltip",
+    default: () => h(OnyxButton, { label: "Hover me to show the tooltip" }),
     open: "hover",
   },
 } satisfies Story;
@@ -67,7 +72,7 @@ export const Hover = {
 export const Click = {
   args: {
     ...Default.args,
-    default: () => "Click me to show the tooltip",
+    default: () => h(OnyxButton, { label: "Click me to show the tooltip" }),
     open: "click",
   },
 } satisfies Story;
@@ -99,5 +104,15 @@ export const Danger = {
   args: {
     ...Default.args,
     color: "danger",
+  },
+} satisfies Story;
+
+/**
+ * This example shows a tooltip that uses the slot for custom content.
+ */
+export const CustomContent = {
+  args: {
+    ...Default.args,
+    tooltip: () => h("div", [h("span", "This is "), h("b", "custom content")]),
   },
 } satisfies Story;