+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # alpha-test-app
 
+## 0.0.1-alpha.52
+
+### Patch Changes
+
+- Updated dependencies [b7758f7]
+  - sit-onyx@1.0.0-alpha.55
+
+## 0.0.1-alpha.51
+
+### Patch Changes
+
+- Updated dependencies [40b72b2]
+  - sit-onyx@1.0.0-alpha.54
+
+## 0.0.1-alpha.50
+
+### Patch Changes
+
+- Updated dependencies [10fb012]
+- Updated dependencies [cb3a72b]
+  - sit-onyx@1.0.0-alpha.53
+
 ## 0.0.1-alpha.49
 
 ### Patch Changes