+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.10
+
+### Patch Changes
+
+- Updated dependencies [b7758f7]
+  - sit-onyx@1.0.0-alpha.55
+
+## 1.0.0-alpha.9
+
+### Patch Changes
+
+- Updated dependencies [40b72b2]
+  - sit-onyx@1.0.0-alpha.54
+
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- ce4945a: prevent "cannot read/set of undefined" errors for unregistered plugins
+
+## 1.0.0-alpha.7
+
+### Patch Changes
+
+- Updated dependencies [10fb012]
+- Updated dependencies [cb3a72b]
+  - sit-onyx@1.0.0-alpha.53
+
 ## 1.0.0-alpha.6
 
 ### Patch Changes