+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,29 @@
 # sit-onyx
 
+## 1.0.0-alpha.55
+
+### Patch Changes
+
+- b7758f7: fix(OnyxTooltip): render newlines in text property
+
+## 1.0.0-alpha.54
+
+### Minor Changes
+
+- 40b72b2: feat(OnyxTooltip): add`tooltip` slot, close tooltip with Escape in "hover" mode
+
+## 1.0.0-alpha.53
+
+### Minor Changes
+
+- cb3a72b: add CSS variables for box shadows
+
+  See [token documentation](https://onyx.schwarz/tokens/shadows.html) for a full list of available shadows.
+
+### Patch Changes
+
+- 10fb012: fix(OnyxCheckbox): force required/optional marker to be visible on label truncation
+
 ## 1.0.0-alpha.52
 
 ### Minor Changes