+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-alpha.55
+
+### Patch Changes
+
+- Updated dependencies [b7758f7]
+  - sit-onyx@1.0.0-alpha.55
+
+## 1.0.0-alpha.54
+
+### Patch Changes
+
+- Updated dependencies [40b72b2]
+  - sit-onyx@1.0.0-alpha.54
+
+## 1.0.0-alpha.53
+
+### Major Changes
+
+- 83192d4: bump minimum vitepress version to 1.0.0
+
+## 1.0.0-alpha.52
+
+### Patch Changes
+
+- Updated dependencies [10fb012]
+- Updated dependencies [cb3a72b]
+  - sit-onyx@1.0.0-alpha.53
+
 ## 1.0.0-alpha.51
 
 ### Patch Changes